Cleanup and renaming
