+++ android/todo-full-sync/app/src/main/java/io/realm/todo/TasksActivity.java
@@ -28,17 +28,24 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.EditText;
+import android.widget.TextView;
 
 import io.realm.Realm;
+import io.realm.RealmResults;
 import io.realm.Sort;
+import io.realm.SyncConfiguration;
 import io.realm.SyncUser;
 import io.realm.todo.model.Item;
 import io.realm.todo.model.Project;
-import io.realm.todo.ui.ItemsRecyclerAdapter;
+import io.realm.todo.ui.TasksRecyclerAdapter;
 
-public class ItemsActivity extends AppCompatActivity {
+public class TasksActivity extends AppCompatActivity {
+
+    public static final String INTENT_EXTRA_PROJECT_URL = "TasksActivity.projectUrl";
 
     private Realm realm;
+    private TextView statusView;
+    private RealmResults<Item> tasks;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -46,31 +53,44 @@ public class ItemsActivity extends AppCompatActivity {
         setContentView(R.layout.activity_items);
 
         setSupportActionBar(findViewById(R.id.toolbar));
+        statusView = findViewById(R.id.status);
 
-        String projectId = getIntent().getStringExtra("project_id");
+        String projectUrl = getIntent().getStringExtra(INTENT_EXTRA_PROJECT_URL);
 
         findViewById(R.id.fab).setOnClickListener(view -> {
             View dialogView = LayoutInflater.from(this).inflate(R.layout.dialog_task, null);
             EditText taskText = dialogView.findViewById(R.id.task);
-            new AlertDialog.Builder(ItemsActivity.this)
+            new AlertDialog.Builder(TasksActivity.this)
                     .setTitle("Add a new task")
                     .setMessage("What do you want to do next?")
                     .setView(dialogView)
                     .setPositiveButton("Add", (dialog, which) -> realm.executeTransactionAsync(realm -> {
                         Item item = new Item();
                         item.setBody(taskText.getText().toString());
-                        realm.where(Project.class).equalTo("projectId", projectId).findFirst().getTasks().add(item);
+                        realm.where(Project.class).findFirst().getTasks().add(item);
                     }))
                     .setNegativeButton("Cancel", null)
                     .create()
                     .show();
         });
 
-        realm = Realm.getDefaultInstance();
-        Project project = realm.where(Project.class).equalTo("projectId", projectId).findFirst();
+        SyncConfiguration config = SyncUser.current()
+                .createConfiguration(projectUrl)
+                .fullSynchronization()
+                .build();
+        realm = Realm.getInstance(config);
+        Project project = realm.where(Project.class).findFirst();
 
         setTitle(project.getName());
-        final ItemsRecyclerAdapter itemsRecyclerAdapter = new ItemsRecyclerAdapter(project.getTasks().sort("timestamp", Sort.ASCENDING));
+        tasks = project.getTasks().where().sort("timestamp", Sort.ASCENDING).findAllAsync();
+        tasks.addChangeListener(tasks -> {
+            if (tasks.isEmpty()) {
+                setStatus("Press + to add a new task");
+            } else {
+                setStatus("");
+            }
+        });
+        final TasksRecyclerAdapter itemsRecyclerAdapter = new TasksRecyclerAdapter(tasks);
         RecyclerView recyclerView = findViewById(R.id.recycler_view);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
         recyclerView.setAdapter(itemsRecyclerAdapter);
@@ -87,8 +107,7 @@ public class ItemsActivity extends AppCompatActivity {
                 int position = viewHolder.getAdapterPosition();
                 String id = itemsRecyclerAdapter.getItem(position).getItemId();
                 realm.executeTransactionAsync(realm -> {
-                    Item item = realm.where(Item.class).equalTo("itemId", id)
-                            .findFirst();
+                    Item item = realm.where(Item.class).equalTo("itemId", id).findFirst();
                     if (item != null) {
                         item.deleteFromRealm();
                     }
@@ -98,6 +117,10 @@ public class ItemsActivity extends AppCompatActivity {
 
         ItemTouchHelper itemTouchHelper = new ItemTouchHelper(simpleItemTouchCallback);
         itemTouchHelper.attachToRecyclerView(recyclerView);
+    }
+
+    private void setStatus(String str) {
+        statusView.setText(str);
     }
 
     @Override
