+++ android/todo-full-sync/app/src/main/res/layout/content_items.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="io.realm.todo.ItemsActivity"
+    tools:context="io.realm.todo.TasksActivity"
     tools:showIn="@layout/activity_items">
 
     <android.support.v7.widget.RecyclerView
