+++ android/todo-query-based-sync/app/src/main/res/layout/task_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="io.realm.todo.ItemsActivity">
+    tools:context="io.realm.todo.TasksActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
